Capital letters in header names in sparse decompositions
