+++ CMakeLists.txt
@@ -209,12 +209,13 @@ set(${PROJECT_NAME}_DECOMPOSITIONS_SPARSE_CHOLMOD_HEADERS
     include/eigenpy/decompositions/sparse/cholmod/CholmodSupernodalLLT.hpp)
 
 set(${PROJECT_NAME}_DECOMPOSITIONS_SPARSE_ACCELERATE_HEADERS
-    include/eigenpy/decompositions/sparse/accelerate/accelerate.hpp)
+    include/eigenpy/decompositions/sparse/accelerate/Accelerate.hpp)
 
 set(${PROJECT_NAME}_DECOMPOSITIONS_SPARSE_HEADERS
-    include/eigenpy/decompositions/sparse/LLT.hpp
-    include/eigenpy/decompositions/sparse/LDLT.hpp
-    include/eigenpy/decompositions/sparse/LU.hpp
+    include/eigenpy/decompositions/sparse/SimplicialLLT.hpp
+    include/eigenpy/decompositions/sparse/SimplicialLDLT.hpp
+    include/eigenpy/decompositions/sparse/SparseLU.hpp
+    include/eigenpy/decompositions/sparse/SparseQR.hpp
     include/eigenpy/decompositions/sparse/SimplicialCholesky.hpp
     include/eigenpy/decompositions/sparse/SparseSolverBase.hpp)
