+++ include/eigenpy/decompositions/sparse/accelerate/Accelerate.hpp
